split all the things.
